adding unique tracks to the playlist

+++ music/src/main/java/by/nyurush/music/controller/command/impl/common/DeletePlaylistCommandImpl.java
@@ -8,12 +8,15 @@ import by.nyurush.music.service.impl.PlaylistService;
 import by.nyurush.music.util.constant.ConstantAttributes;
 import by.nyurush.music.util.constant.ConstantMessages;
 import by.nyurush.music.util.constant.ConstantPathPages;
+import by.nyurush.music.util.language.ResourceBundleUtil;
+import by.nyurush.music.util.validation.StringUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.Optional;
+import java.util.ResourceBundle;
 
 import static by.nyurush.music.util.constant.ConstantAttributes.ERROR_MESSAGE;
 import static by.nyurush.music.util.constant.ConstantAttributes.SUCCESS_MESSAGE;
@@ -23,16 +26,23 @@ public class DeletePlaylistCommandImpl implements Command {
 
     @Override
     public CommandResult execute(HttpServletRequest req, HttpServletResponse resp) throws ServiceException {
+        ResourceBundle rb = ResourceBundleUtil.getResourceBundle(req);
         PlaylistService playlistService = new PlaylistService();
-        Integer id = Integer.parseInt(req.getParameter(ConstantAttributes.PLAYLIST_ID));
+        String idStr = req.getParameter(ConstantAttributes.PLAYLIST_ID);
+        if(StringUtil.isNullOrEmpty(idStr)) {
+            req.setAttribute(ERROR_MESSAGE, rb.getString(ConstantMessages.INVALID_FIND_SONG));
+            return CommandResult.forward(ConstantPathPages.PATH_PAGE_PLAYLISTS);
+        }
+        Integer id = Integer.parseInt(idStr);
         Optional<Playlist> playlist = playlistService.findById(id);
         if (playlist.isPresent()) {
             playlistService.delete(playlist.get());
         } else {
             LOGGER.warn("Playlist was not found");
-            req.setAttribute(ERROR_MESSAGE, ConstantMessages.INVALID_FIND_SONG);
+            req.setAttribute(ERROR_MESSAGE, rb.getString(ConstantMessages.INVALID_FIND_SONG));
+            return CommandResult.forward(ConstantPathPages.PATH_PAGE_PLAYLISTS);
         }
-        req.setAttribute(SUCCESS_MESSAGE, ConstantMessages.SUCCESSFUL_DELETE_PLAYLIST);
+        req.setAttribute(SUCCESS_MESSAGE, rb.getString(ConstantMessages.SUCCESSFUL_DELETE_PLAYLIST));
         return CommandResult.forward(ConstantPathPages.PATH_PAGE_PLAYLISTS);
     }
 }
